Move collection from Refresher to Inventory::Collector
