+++ app/models/manageiq/providers/kubernetes/inventory/collector.rb
@@ -1,2 +1,3 @@
 class ManageIQ::Providers::Kubernetes::Inventory::Collector < ManageIQ::Providers::Inventory::Collector
+  require_nested :ContainerManager
 end
